feat: add a little refactor to improve the module organization on the project
